+++ CHANGELOG.md
@@ -9,8 +9,6 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
-- bumped pytz to newest version >=2025
-
 ### Changes
 
 - don't require env settings for ldap and wiki tests
@@ -21,9 +19,18 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Fixed
 
-- fixed temporal entity tests for newest pytz version
-
 ### Security
+
+## [1.5.0] - 2025-09-30
+
+### Added
+
+- bumped pytz to newest version >=2025
+- BREAKING: add RKI organization as affiliation to ldap persons
+
+### Fixed
+
+- fixed temporal entity tests for newest pytz version
 
 ## [1.4.0] - 2025-09-11
 
@@ -193,6 +200,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - added MEX_BACKEND_API_PARALLELIZATION and MEX_BACKEND_API_CHUNK_SIZE settings
 - added support for sending batches of data to the backend in parallel
+
 
 ### Changes